fix: resolve Vercel build errors (types, Suspense, unused props)
- Type household query results in blueprint and dashboard pages
- Partner complete: pass token into run(t) for type safety; wrap useSearchParams in Suspense
- RitualTransferSummary/SeedDialog: prefix unused props with underscore
